move `relevantFor` to ops
