+++ core/src/models/Destination.ts
@@ -778,30 +778,4 @@ export class Destination extends LoggedModel<Destination> {
       where: { destinationId: instance.id },
     });
   }
-
-  /**
-   * Determine which destinations are interested in this record due to the groups they are tracking
-   */
-  static async relevantFor(
-    record: GrouparooRecord,
-    oldGroups: Group[] = [],
-    newGroups: Group[] = []
-  ) {
-    const combinedGroupIds = [...oldGroups, ...newGroups].map((g) => g.id);
-    const relevantDestinations =
-      combinedGroupIds.length > 0
-        ? await Destination.findAll({
-            where: {
-              [Op.or]: [
-                { collection: "model", modelId: record.modelId },
-                { collection: "group", groupId: { [Op.in]: combinedGroupIds } },
-              ],
-            },
-          })
-        : await Destination.findAll({
-            where: { collection: "model", modelId: record.modelId },
-          });
-
-    return relevantDestinations;
-  }
 }
